requirements.txt
fastapi
uvicorn[standard]
sqlmodel
python-dotenv
pydantic-settings
passlib[bcrypt]
python-jose[cryptography]
python-multipart
websockets
cryptography
pymysql
psycopg2-binary
google-generativeai
google-auth
google-auth-oauthlib